% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grouping.R
\name{addPartition}
\alias{addPartition}
\title{Add grouping partition}
\usage{
addPartition(tree = NULL, grouping, applyTo = "leaf", column = NULL,
  data = NULL)
}
\arguments{
\item{tree}{The tree node to which partition should be added.}

\item{grouping}{The grouping needed to be added. Possible values: all, primary3, greaterThan, coloumn}

\item{applyTo}{Whether to add new node to the lead or to the root.}

\item{column}{The column to split if value to Grouping parameter is column.}

\item{data}{The data to do grouping}
}
\value{
a tree with new node attached to the input
}
\description{
Add a grouping partition either at the leaf or root.
}
\examples{
\dontrun{
allGroups <- addPartition(grouping = "all")
}
}
\author{
thiloshon <thiloshon@gmail.com>
}
