% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grouping.R
\name{addPartitionToLeaf}
\alias{addPartitionToLeaf}
\title{Utility function to add node to the leaf}
\usage{
addPartitionToLeaf(node, value = NULL)
}
\arguments{
\item{node}{The tree node to which new node should be added.}

\item{value}{The node or character to appended to the tree.}
}
\description{
Utility function to add node to the leaf
}
\author{
thiloshon <thiloshon@gmail.com>
}
