% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.biodiv}
\alias{coef.biodiv}
\title{Extract fixed coefficients of biodiversity model object}
\usage{
\method{coef}{biodiv}(object, ...)
}
\arguments{
\item{object}{biodiv fitted object (result from biodiv)}

\item{...}{extra arguments}
}
\description{
Extract fixed coefficients of biodiversity model object
}
\details{
...
}
