% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCleaning.R
\name{coordinate_negated_flag}
\alias{coordinate_negated_flag}
\title{Flag records with coordinates to be corrected to match country}
\usage{
coordinate_negated_flag(gbif_data)
}
\arguments{
\item{gbif_data}{Dataframe from GBIF with two mandatory fields; "decimalLatitude", "decimalLongitude"}
}
\value{
Same dataframe with one additional column; coordinateNegatedFlag
}
\description{
Supplied geographic coordinates have to be transposed or the sign reversed (negated) to place the record in the supplied country
}
\examples{
\dontrun{
dat <- rgbif::occ_data(scientificName = 'Ursus americanus')
flagged_dat <- coordinate_negated_flag(dat$data)
}
}
\author{
thiloshon <thiloshon@gmail.com>
}
