% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCleaning.R
\name{coordinates_decimal_mismatch}
\alias{coordinates_decimal_mismatch}
\title{Flag coordinates with decimal points mismatch.}
\usage{
coordinates_decimal_mismatch(gbif_dataFrame)
}
\arguments{
\item{gbif_data}{Dataframe from GBIF with two mandatory fields; decimalLatitude, decimalLongitude}
}
\value{
Same dataframe with one additional column; decimalPointDifference
}
\description{
Runs quality check of checking coordinates with more than 3 decimal point difference.
}
\details{
The function runs a quality check on coordinates of GBIF data to check if latitude and longitude have
different number of decimal points. If coordinate was recorded by an accepted protocol, its unlikely to
identify both coordinate with varying precision.
}
\examples{
dat <- rgbif::occ_data(scientificName = 'Ursus americanus')
flagged_dat <- coordinates_decimal_mismatch(dat)
}
\author{
thiloshon <thiloshon@gmail.com>
}
