% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCleaning.R
\name{country_code_unknown_flag}
\alias{country_code_unknown_flag}
\title{Flag records with unidentified country codes / country}
\usage{
country_code_unknown_flag(gbif_data)
}
\arguments{
\item{gbif_data}{Dataframe from GBIF with one mandatory field; countryCode}
}
\value{
Same dataframe with one additional column; countryCodeUnknownFlag
}
\description{
Runs quality check of checking if country code / country is recognized.
}
\examples{
dat <- rgbif::occ_data(scientificName = 'Ursus americanus')
flagged_dat <- country_code_unknown_flag(dat$data)
}
\author{
thiloshon <thiloshon@gmail.com>
}
