% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCleaning.R
\name{county_coordinate_mismatch_flag}
\alias{county_coordinate_mismatch_flag}
\title{Flag records with contradicting county to the coordinates}
\usage{
county_coordinate_mismatch_flag(gbif_data)
}
\arguments{
\item{gbif_data}{Dataframe from GBIF with three mandatory field; decimalLatitude, decimalLongitude, county}
}
\value{
Same dataframe with one additional column; countyCoordinateMismatchFlag
}
\description{
Runs quality check of checking if county field and coordinates point same location.
}
\examples{
dat <- rgbif::occ_data(scientificName = 'Ursus americanus')
flagged_dat <- county_coordinate_mismatch_flag(dat$data)
}
\author{
thiloshon <thiloshon@gmail.com>
}
