% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCleaning.R
\name{depth_out_of_range_flag}
\alias{depth_out_of_range_flag}
\title{Flag records with incorrect depth}
\usage{
depth_out_of_range_flag(gbif_data)
}
\arguments{
\item{gbif_data}{Dataframe from GBIF with one mandatory field; depth}
}
\value{
Same dataframe with one additional column; depthOutofRangeFlag
}
\description{
Runs quality check of checking if depth is wthin possible range.
}
\details{
The depth (dwc:depth) should be less than zero (0) or maximum depth is greater than 11,000 meters
}
\examples{
dat <- rgbif::occ_data(scientificName = 'Ursus americanus')
flagged_dat <- uncertainty_outofrange_flag(dat$data)
}
\author{
thiloshon <thiloshon@gmail.com>
}
