% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{devi}
\alias{devi}
\title{Deviance fraction of bioviv output}
\usage{
devi(fit, ...)
}
\arguments{
\item{fit}{as returned from the biodiv function.}
}
\value{
deviance fraction (see details)
}
\description{
Deviance fraction of bioviv output
}
\details{
Deviance fraction calculated by comparing the deviance between the estimated model (M) and the saturated model (S) to the deviance between the constant model(C) and the saturated model (S). The constant and saturated models are optimized with same overdispersion parameter as the estimated model. The deviance fraction is defined as: 1-(logLik(M)-logLik(S))/(logLik(C)-logLik(S))
}
