% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemporalCleaning.R
\name{first_of_year_flag}
\alias{first_of_year_flag}
\title{Flag records with dates set to first day of the year}
\usage{
first_of_year_flag(GBIF_Data)
}
\arguments{
\item{gbif_data}{Dataframe from GBIF with two mandatory fields; month and day.}
}
\value{
Same dataframe with one column; firstOfYearFlag.
}
\description{
Event date (month, day) is first of year eventDate=1978-01-01, or year=1978, month=1, day=1.
}
\details{
The function runs a quality check to find if the date of occurance is set to the first day of the year.
This might mean the date was not recorded and ehile digitization the date was automatically set my system as the first day.
}
\examples{
dat <- rgbif::occ_data(scientificName = 'Ursus americanus')
flagged_dat <- first_of_year_flag(dat$data)
}
\author{
thiloshon <thiloshon@gmail.com>
}
