% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCleaningSet02.R
\name{format_checking}
\alias{format_checking}
\title{Internal function to check validity of input GBIF dataset}
\usage{
format_checking(gbif_data, variable_vector = NULL)
}
\arguments{
\item{gbif_data}{Object to check validity.}

\item{variable_vector}{Vector containing required filed names.}
}
\value{
Same object stripped upto dataframe
}
\description{
Finds is input object is one of these, dataframe, gbif_data, dwca_gbif. And also checks if
the needed columns are present in the dataset.
}
\author{
thiloshon <thiloshon@gmail.com>
}
