% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCleaningSet02.R
\name{gbif_issues_flag}
\alias{gbif_issues_flag}
\title{Flag records with GBIF Issues}
\usage{
gbif_issues_flag(gbif_data)
}
\arguments{
\item{gbif_data}{Dataframe from GBIF with one mandatory field; issue.}
}
\value{
Same dataframe with four additional columns; countryDerivedFromCoordinatesFlag, geodeticDatumConvertedFlag,
geodeticDatumInvalidFlag, geodeticDatumAssumedFlag.
}
\description{
Finds records with issues field filled.
}
\examples{
\dontrun{
dat <- rgbif::occ_data(scientificName = 'Ursus americanus')
flagged_dat <- gbif_issues_flag(dat$data)
}
}
\author{
thiloshon <thiloshon@gmail.com>
}
