% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCleaning.R
\name{georeference_protocol_flag}
\alias{georeference_protocol_flag}
\title{Flag questionable georeference protocol records}
\usage{
georeference_protocol_flag(gbif_data)
}
\arguments{
\item{gbif_data}{Dataframe from GBIF with one mandatory field; georeferenceProtocol}
}
\value{
Same dataframe with one additional column; georeferenceProtocolFlag
}
\description{
Runs quality check of checking authenticity of georeference protocols and flags accordingly.
}
\details{
The function runs a quality check on georeference protocol of GBIF data.
Checks if the georeferenceProtocol is valid. Generally, coordinates determined from GPS have an uncertainty of a few metres,
especially if the Horizontal Dilution of Precision (HDOP) is also recorded. Coordinates determined from modern satellite maps,
for example Google Maps/Earth are also quite precise (although beware of satellite image rectification issues). On the other hand
coordinates determined from say 1:1 million maps, or from gazetteers accurate to one minute, are much less precise.
}
\examples{
dat <- rgbif::occ_data(scientificName = 'Ursus americanus')
flagged_dat <- georeference_protocol_flag(dat$data)
}
\author{
thiloshon <thiloshon@gmail.com>
}
