% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCleaning.R
\name{georeference_verification_status_flag}
\alias{georeference_verification_status_flag}
\title{Flag unverified georeferenced records}
\usage{
georeference_verification_status_flag(gbif_data)
}
\arguments{
\item{gbif_data}{Dataframe from GBIF with one mandatory field; georeferenceVerificationStatus}
}
\value{
Same dataframe with one additional column; georeferenceVerificationStatusFlag
}
\description{
Runs quality check of checking verification of georeferences.
}
\details{
The function runs a quality check on georeference protocol of GBIF data.
Checks if the georeferenceVerificationStatus is valid.
Where an occurrence record is on the edge of its range, or is a range extension, then it is useful to know whether the
location coordinates have been verified
}
\examples{
dat <- rgbif::occ_data(scientificName = 'Ursus americanus')
flagged_dat <- georeference_verification_status_flag(dat$data)
}
\author{
thiloshon <thiloshon@gmail.com>
}
