% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grouping.R
\name{getPartitions}
\alias{getPartitions}
\title{Utility function to return all partition paths}
\usage{
getPartitions(tree)
}
\arguments{
\item{tree}{The tree node to which nodes should be returned.}
}
\description{
Utility function to return all partition paths
}
\author{
thiloshon <thiloshon@gmail.com>
}
