% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemporalCleaning.R
\name{identified_pre_event_flag}
\alias{identified_pre_event_flag}
\title{Flag records with incorrect event date, identification date pair.}
\usage{
identified_pre_event_flag(GBIF_Data)
}
\arguments{
\item{gbif_data}{Dataframe from GBIF with two mandatory fields; dateIdentified and eventDate.}
}
\value{
Same dataframe with two column; identifiedPreEventFlag, identifiedPreEventDiff
}
\description{
The date of identification (dwc:dateIdentified) is earlier than the event date dateIdentified < eventDate
}
\details{
The function runs a quality check to find if the date of identification is before date of occurance.
This is not possible at all as identification can only be done on or after occurance.
}
\examples{
dat <- rgbif::occ_data(scientificName = 'Ursus americanus')
flagged_dat <- identified_pre_event_flag(dat$data)
}
\author{
thiloshon <thiloshon@gmail.com>
}
