% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemporalCleaning.R
\name{impropable_identified_date_flag}
\alias{impropable_identified_date_flag}
\title{Flag records with incorrect identification date.}
\usage{
impropable_identified_date_flag(GBIF_Data)
}
\arguments{
\item{gbif_data}{Dataframe from GBIF with two mandatory fields; dateIdentified and eventDate.}
}
\value{
Same dataframe with one column; dateIdentified
}
\description{
The date of identification (dwc:dateIdentified) falls prior to Linnaeus (1753) or after
the current date. dateIdentified < 1753 or in future
}
\examples{
dat <- rgbif::occ_data(scientificName = 'Ursus americanus')
flagged_dat <- impropable_identified_date_flag(dat$data)
}
\author{
thiloshon <thiloshon@gmail.com>
}
