% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCleaningSet02.R
\name{invasive_flags}
\alias{invasive_flags}
\title{Flag records with species that are invasive.}
\usage{
invasive_flags(gbif_data)
}
\arguments{
\item{gbif_data}{Dataframe from GBIF with one mandatory fields; scientificName}
}
\value{
Same dataframe with one additional column; invasiveFlags
}
\description{
It is important to know if an occurrence is natural, rather than an escapee from captivity, or say, a plant cultivated in a park,
or indeed if the occurrence is extralimital to is normal range, e.g. a vagrant migratory bird that has drifted way off-course.
Usually these records are excluded from spatial analyses.
}
\examples{
dat <- rgbif::occ_data(scientificName = 'Ursus americanus')
flagged_dat <- invasive_flags(dat$data)
}
\author{
thiloshon <thiloshon@gmail.com>
}
