% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grouping.R
\name{listPartitions}
\alias{listPartitions}
\title{View all partitions added}
\usage{
listPartitions(tree, onlyLeaf = FALSE)
}
\arguments{
\item{tree}{The tree node to which partition should be listed.}

\item{onlyLeaf}{List only the leaf partitions. i.e Only the final outputs.}
}
\value{
a vector with all possible parttions
}
\description{
View all possible partitions of the node with path.
}
\examples{
\dontrun{
allGroups <- addPartition(grouping = "all")
listPartitions(allGroups)
}
}
\author{
thiloshon <thiloshon@gmail.com>
}
