% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCleaning.R
\name{locality_coordinate_mismatch_flag}
\alias{locality_coordinate_mismatch_flag}
\title{Flag records with contradicting locality to the coordinates}
\usage{
locality_coordinate_mismatch_flag(gbif_data)
}
\arguments{
\item{gbif_data}{Dataframe from GBIF with three mandatory field; decimalLatitude, decimalLongitude, locality}
}
\value{
Same dataframe with two additional columns; localityCoordinateMismatchFlag, generatedLocalityCoordinate
}
\description{
Runs quality check of checking if locality field and coordinates point same location.
}
\examples{
dat <- rgbif::occ_data(scientificName = 'Ursus americanus')
flagged_dat <- locality_coordinate_mismatch_flag(dat$data)
}
