% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.biodiv}
\alias{logLik.biodiv}
\title{Log likelihood of biodiv object}
\usage{
\method{logLik}{biodiv}(object, ...)
}
\arguments{
\item{object}{biodiv fitted object (result from biodiv)}

\item{...}{extra arguments}
}
\description{
Log likelihood of biodiv object
}
\details{
...
}
