% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCleaningSet02.R
\name{native_flags}
\alias{native_flags}
\title{Flag records with species that are not native to a geographical area.}
\usage{
native_flags(gbif_data)
}
\arguments{
\item{gbif_data}{Dataframe from GBIF with one mandatory fields; scientificName}
}
\value{
Same dataframe with three additional columns; nativeFlags, isIntroduced, isCultivated
}
\description{
It is important to know if an occurrence is natural, rather than an escapee from captivity, or say, a plant cultivated in a park,
or indeed if the occurrence is extralimital to is normal range, e.g. a vagrant migratory bird that has drifted way off-course.
Usually these records are excluded from spatial analyses.
}
\examples{
dat <- rgbif::occ_data(scientificName = 'Ursus americanus')
flagged_dat <- native_flags(dat$data)
}
\author{
thiloshon <thiloshon@gmail.com>
}
