% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.biodiv}
\alias{plot.biodiv}
\title{Plot biodiv object}
\usage{
\method{plot}{biodiv}(x, ...)
}
\arguments{
\item{x}{object returned from biodiv}

\item{...}{extra arguments}
}
\description{
Plot biodiv object
}
\details{
...
}
