% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.biodivres}
\alias{plot.biodivres}
\title{Plot biodivres object}
\usage{
\method{plot}{biodivres}(x, ...)
}
\arguments{
\item{x}{residuals from a biodiv model}

\item{...}{extra arguments}
}
\description{
Plot biodivres object
}
\details{
...
}
