% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCleaning.R
\name{precision_uncertainty_mismatch_flag}
\alias{precision_uncertainty_mismatch_flag}
\title{Flag records with swapped precision and uncertainty}
\usage{
precision_uncertainty_mismatch_flag(gbif_data)
}
\arguments{
\item{gbif_data}{Dataframe from GBIF with one mandatory field; coordinatePrecision}
}
\value{
Same dataframe with one additional column; precisionUncertaintyMismatchFlag
}
\description{
CoordinateUncertaintyInMeters and coordinatePrecision appear swapped as precision is integer > 0 and uncertainty is 0-<=1
}
\examples{
\dontrun{
dat <- rgbif::occ_data(scientificName = 'Ursus americanus')
flagged_dat <- precision_uncertainty_mismatch_flag(dat$data)
}
}
\author{
thiloshon <thiloshon@gmail.com>
}
