% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.biodiv}
\alias{print.biodiv}
\title{Print biodiv object}
\usage{
\method{print}{biodiv}(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{extra arguments}
}
\description{
Print biodiv object
}
\details{
...
}
