% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.biodivcoef}
\alias{print.biodivcoef}
\title{Print biodivcoef object}
\usage{
\method{print}{biodivcoef}(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{extra arguments}
}
\description{
Print biodivcoef object
}
\details{
...
}
