% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{R2}
\alias{R2}
\title{R2 of bioviv output}
\usage{
R2(fit, ...)
}
\arguments{
\item{fit}{as returned from the biodiv function}
}
\value{
R2
}
\description{
R2 of bioviv output
}
\details{
Calculates R-squared coefficient of determination as: 1-SS(model residuals)/SS(total residuals)
}
