% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_function.R, R/spatialCleaning.R
\name{repeating_digits}
\alias{repeating_digits}
\alias{repeating_digits}
\title{Flag repeating digits}
\usage{
repeating_digits(gbif_data)

repeating_digits(gbif_data)
}
\arguments{
\item{gbif_data}{Dataframe from GBIF with two mandatory fields; decimalLatitude and decimalLongitude.}

\item{gbif_data}{Dataframe from GBIF with two mandatory fields; decimalLatitude and decimalLongitude.}
}
\value{
Same dataframe with two additional columns; latRepeatCount and longRepeatCount. Both shows the number of digits that are repeating

Same dataframe with two additional columns; latRepeatCount and longRepeatCount. Both shows the number of digits that are repeating
}
\description{
Runs quality check of finding repeated digits and flags accordingly.

Runs quality check of finding repeated digits and flags accordingly.
}
\details{
The function runs a quality check on two fields of GBIF data - decimalLatitude and decimalLongitude.
Checks if the decimal values have repeated values at the end of the value. Repeated values might
mean error in digitisation or conversions of records

The function runs a quality check on two fields of GBIF data - decimalLatitude and decimalLongitude.
Checks if the decimal values have repeated values at the end of the value. Repeated values might
mean error in digitisation or conversions of records
}
\examples{
dat <- rgbif::occ_data(scientificName = 'Ursus americanus')
flagged_dat <- repeating_digits(dat)
dat <- rgbif::occ_data(scientificName = 'Ursus americanus')
flagged_dat <- repeating_digits(dat)
}
\author{
thiloshon <thiloshon@gmail.com>

thiloshon <thiloshon@gmail.com>
}
