% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.biodiv}
\alias{residuals.biodiv}
\title{Extract residuals from biodiv object}
\usage{
\method{residuals}{biodiv}(object, ...)
}
\arguments{
\item{object}{biodiv fitted object (result from biodiv)}

\item{...}{extra arguments for TMB's oneStepPredict}
}
\description{
Extract residuals from biodiv object
}
\details{
...
}
