% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_taxonrank.R
\name{resolve_taxonrank}
\alias{resolve_taxonrank}
\title{Resolve taxonranks}
\usage{
resolve_taxonrank(GBIF_Data, gnr_score = 0.75, tidy_output = TRUE,
  resolve = dim(GBIF_Data)[1], location = "Australia", zoom = 4,
  simplify = FALSE, upto = "subspecies", ...)
}
\arguments{
\item{GBIF_Data}{Dataframe from GBIF or object of type gbif.}

\item{tidy_output}{When the output is very long, whether to cut off unimportant parts}

\item{resolve}{The amount of names to resolve. Default is all. If number of names is high, can use this.}

\item{location}{The location to consider when plotting maps for the suggested names}

\item{zoom}{The zoom value to consider when plotting the worldmap centered at the location parameter. 4 means city and so on...}

\item{simplify}{Which workflow to follow. simplyfied workflow runs faster than the other but is less accurate.}

\item{upto}{the rank upto which function should try to resolve}

\item{gnr_resolve}{The score to consider as qualifier in the gnr_resolve internal function}
}
\value{
A list with multiple outputs.
}
\description{
Resolve taxonranks which are not identified at the Species, Subspecies or Variety rank.
}
\details{
This function resolves the records which are not identified at the specie, subscpecie or variety level.
It checks for similar records from other sources and tries to build series of world maps with original and suggested species distributions.
The simple version tries to do the same but by taking synonyms and other names into account
}
\seealso{
[resolve_taxon_inspect()]
}
