% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolveCoordinates.R
\name{resolveCoordinates}
\alias{resolveCoordinates}
\title{Resolve coordinates}
\usage{
resolveCoordinates(gbif_data)
}
\arguments{
\item{gbif_data}{Dataframe from GBIF or object of type gbif.}
}
\value{
The original dataframe with added columns.
}
\description{
Resolve coordinates which are not pointed at the correct country.
}
\seealso{
[resolve_taxon_inspect()]
}
