% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_check.R
\name{runQuickCheck}
\alias{runQuickCheck}
\title{Run a questionnaire to get user preferences.}
\usage{
runQuickCheck()
}
\description{
Runs a questionnaire to get the user preferences to use in all the quality checks.
}
\author{
thiloshon <thiloshon@gmail.com>
}
