% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{runwithout}
\alias{runwithout}
\title{Re-run a model without a given covariate}
\usage{
runwithout(fit, what)
}
\arguments{
\item{fit}{as returned from the biodiv function.}

\item{what}{quoted name of covariate to exclude (done by assigning it to its average value)}
}
\value{
The fraction that the average variance is reduced by including the named covariate
}
\description{
Re-run a model without a given covariate
}
