% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCleaning.R
\name{stateProvinceCoordinateMismatchFlag}
\alias{stateProvinceCoordinateMismatchFlag}
\title{Flag records with contradicting stateProvince to the coordinates}
\usage{
stateProvinceCoordinateMismatchFlag(gbif_data)
}
\arguments{
\item{gbif_data}{Dataframe from GBIF with three mandatory field; decimalLatitude, decimalLongitude, stateProvince}
}
\value{
Same dataframe with one additional column; stateProvinceCoordinateMismatchFlag
}
\description{
Runs quality check of checking if stateProvince field and coordinates point same location.
}
\examples{
dat <- rgbif::occ_data(scientificName = 'Ursus americanus')
flagged_dat <- stateProvinceCoordinateMismatchFlag(dat$data)
}
\author{
thiloshon <thiloshon@gmail.com>
}
