% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCleaning.R
\name{uncertainty_outofrange_flag}
\alias{uncertainty_outofrange_flag}
\title{Flag records with incorrect coordinateuncertainity}
\usage{
uncertainty_outofrange_flag(gbif_data)
}
\arguments{
\item{gbif_data}{Dataframe from GBIF with one mandatory field; coordinateUncertaintyInMeters}
}
\value{
Same dataframe with one additional column; uncertaintyOutofRangeFlag
}
\description{
Runs quality check of checking if coordinateUncertaintyInMeters is wthin possible range.
}
\details{
The Geopoint uncertainty (dwc:coordinateUncertaintyInMeters) should be a whole number and greater than zero (meters)
coordinateUncertaintyInMeters=integer<0
}
\examples{
dat <- rgbif::occ_data(scientificName = 'Ursus americanus')
flagged_dat <- uncertainty_outofrange_flag(dat$data)
}
\author{
thiloshon <thiloshon@gmail.com>
}
