% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{varianceReduction}
\alias{varianceReduction}
\title{Variance reduction by each term}
\usage{
varianceReduction(fit, lookat = c("siz", "temp", "asampl", "mesh", "density",
  "npp", "depth", "abund", "lat", "lon"))
}
\arguments{
\item{fit}{as returned from the biodiv function.}

\item{lookat}{a vector of quoted covariate names to be looked at}
}
\value{
A vector of variance reduction fractions
}
\description{
Variance reduction by each term
}
\examples{
order<-c(2,4,1,3)
fits <- lapply(order, function(i)biodiv(species,i))
cols <- lapply(fits, function(fit){vr<-varianceReduction(fit); d<-devi(fit); c(vr/sum(vr,na.rm=TRUE)*d,unexplained=1-d)})
bars <- do.call(cbind, cols)
colnames(bars) <- c("Neutral", "Best", "Latitude", "Metabolic")[order]
par(mar=c(3,4,1,9)+0.5,oma=c(0,0,0,3), ask=FALSE)
colvec <- rgb(rbind(c(105,0,0),c(204,0,0),c(204,98,0),c(204,196,0),c(140,204,140),
          c(0,204,130),c(0,131,204),c(65,0,204),c(163,0,204),
          c(120,120,120)),alpha=204,max=204)
op <- palette(colvec)
barplot(bars,col=c(op,"white"),names.arg=colnames(bars),
      legend.text=rownames(bars),ylab="Proportion of deviance",
        args.legend=list(x=7.5,y=1.05,bty="n"))

}
