% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{biodiv}
\alias{biodiv}
\title{Fit biodiversity model}
\usage{
biodiv(data, conf, fixK = NULL, useTotCatch = TRUE, useAswept = TRUE,
  run = TRUE, ...)
}
\arguments{
\item{data}{A data set in same format as the species data set supplied with this package}

\item{conf}{A code identifying the model configuration (-2=constant model, -1=saturated model, 1=Neutral, 2=Best, 3=Latitude, and 4=Metabolic)}

\item{fixK}{optionally a value used to fix the k parameter.}

\item{run}{if FALSE return AD object without running the optimization}

\item{...}{extra arguments to MakeADFun}
}
\value{
an object of class \code{biodiv}
}
\description{
Fit biodiversity model
}
\details{
The model ...
}
\examples{
data(species)
fit <- biodiv(species, conf=1)
}
