% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodiversity.R
\name{biodiversity}
\alias{biodiversity}
\title{Launch biodiversity with default browser}
\usage{
biodiversity()
}
\value{
shiny webpage
}
\description{
Launch biodiversity with default browser
}
\examples{
ShinyApp <-  1
\dontrun{
biodiversity()
}

}
